# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BaseResponse:
    properties:
      data: {}
      invalidParams: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.RenewRequest:
    properties:
      refresh_token:
        type: string
    type: object
  dto.SignUpRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
host: localhost:5005
info:
  contact: {}
  description: This is a server for the food delivery service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Food Delivery API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and generates access and refresh tokens.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Invalid login request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: User login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using a refresh token.
      parameters:
      - description: Refresh Token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RenewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/set:
    get:
      consumes:
      - application/json
      description: Sets the Authorization header with a Bearer token.
      parameters:
      - description: Authorization Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set Authorization Header
      tags:
      - Auth
  /auth/sign:
    post:
      consumes:
      - application/json
      description: Registers a new user and generates access and refresh tokens.
      parameters:
      - description: User sign-up details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sign-up successful
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Invalid sign-up request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: User sign up
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: JWT security accessToken. Please add it in the format "Bearer {AccessToken}"
      to authorize your requests.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"